nearly finished
worked on the prompt parsing mostly, should be able to get this done in a day! ^_^
committing before i do some changes to how i parsed it,though
the writing shoulddd take a short bit of time? and it's barely even writing. just kind of a loose setting with things to do and character(s) to talk to