sdl has been added forreal, epiccontext is epic
now refactoring the code to work in real time using function returns and timing things based off that, rather than relying on pausing the thread to delay actions.
while that works well in the console, it does not work well in a program with a main render loop, like sdl/opengl-based projects
also got rid of the circular dependencies by making epiccontext be completely separate from everything else, only containing data that could then be manipulated from the other classes. this could be done by using an Epic instance with an EpicContext instance to get the current node, using the current node string of the context to grab the corresponding node in the Epic instance